+++ packages/core/src/State.ts
@@ -20,11 +20,17 @@ type GraphProperties = {
 }
 
 function getConfigHome (): string {
+  // Even though Windows and MacOS do not follow the XDG Base Directory spec we
+  // allow the environment variable XDG_CONFIG_HOME to override the platform's
+  // default in case users want to maintain a consistent configuration setup
+  // across multiple platforms.
   switch (process.platform) {
     case 'win32':
-      return process.env.APPDATA || path.join(os.homedir(), 'AppData', 'Roaming')
+      return process.env.XDG_CONFIG_HOME || process.env.APPDATA ||
+        path.join(os.homedir(), 'AppData', 'Roaming')
     case 'darwin':
-      return path.join(os.homedir(), 'Library', 'Application Support')
+      return process.env.XDG_CONFIG_HOME ||
+        path.join(os.homedir(), 'Library', 'Application Support')
     default:
       return process.env.XDG_CONFIG_HOME || path.join(os.homedir(), '.config')
   }